fix: corrigir deletar usuário

+++ src/repositories/userRepository.ts
@@ -1,4 +1,4 @@
-import { User } from '@prisma/client';
+import { User, Participants } from '@prisma/client';
 import { prisma } from '../../prisma/client';
 
 class UserRepository {
@@ -16,10 +16,24 @@ class UserRepository {
     }
 
     public async deleteUser(id: number): Promise<User> {
-        return await prisma.user.delete({
-            where: { id }
+        const user = await prisma.user.findUnique({
+          where: { id },
+          include: { participants: true },
         });
-    }
+      
+        if (user) {
+          await prisma.participants.deleteMany({
+            where: { userEmail: user.email },
+          });
+      
+          return await prisma.user.delete({
+            where: { id },
+          });
+        } else {
+          throw new Error("Usuário não encontrado");
+        }
+      }
+      
 
     public async findAllUsers() {
